+++ src/components/DedicationsSection.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+function DedicationsSection() {
+  return (
+    <span className="text-white">
+      Dedicated to our coordinator Angelique and countless turtle watch
+      volunteers!
+    </span>
+  );
+}
+
+export default DedicationsSection;