+++ src/components/EditionSection.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+function EditionSection() {
+  return (
+    <>
+      <img src="./static/images/editions/easter.svg" alt="" width={200} />
+      <span className="text-white">
+        Happy Easter to you and your loved ones!
+      </span>
+    </>
+  );
+}
+
+export default EditionSection;